+++ AzDevOps/core/projects/public/GetProject.ps1
@@ -1,0 +1,45 @@
+function Get-Project
+{
+ [CmdletBinding(
+  HelpURI = 'https://github.com/Azure-Devops-PowerShell-Module/projects/blob/master/docs/Get-AdoProject.md#get-adoproject',
+  PositionalBinding = $true)]
+ [OutputType([Object])]
+ param (
+  [Parameter(Mandatory = $false)]
+  [Guid]$ProjectId,
+  [Parameter(Mandatory = $false)]
+  [ValidateSet('5.1', '7.1-preview.4')]
+  [string]$ApiVersion = '7.1-preview.4'
+ )
+ begin
+ {
+  Write-Verbose "GetProject    : Begin Processing";
+  Write-Verbose " ProjectId    : $($ProjectId)";
+  Write-Verbose " ApiVersion   : $($ApiVersion)";
+  try
+  {
+   $ErrorActionPreference = 'Stop';
+   $Error.Clear();
+   #
+   # Are we connected
+   #
+   if ($Global:azDevOpsConnected)
+   {
+    if ($ProjectId)
+    {
+     $Uri = $Global:azDevOpsOrg + "_apis/projects/$($ProjectId)?api-version=$($ApiVersion)";
+     return (Invoke-AdoEndpoint -Uri ([System.Uri]::new($Uri)) -Method Get -Headers $Global:azDevOpsHeader -Verbose:$VerbosePreference);
+    }
+    else
+    {
+     $Uri = $Global:azDevOpsOrg + "_apis/projects?api-version=$($ApiVersion)";
+     return (Invoke-AdoEndpoint -Uri ([System.Uri]::new($Uri)) -Method Get -Headers $Global:azDevOpsHeader -Verbose:$VerbosePreference).Value;
+    }
+   }
+  }
+  catch
+  {
+   throw $_;
+  }
+ }
+}